refactor(gfql/polars): fix 2 rename stragglers (new engine_polars refs on feat)
get_degrees DAG dispatch (call/executor.py) and the _cmp_expr predicate unit-test import
were ADDED on the followups branch after the base, so the lazy/engine/polars home-move's
import rewrite didn't cover them. Repoint both to lazy.engine.polars.
